Experimental moving of splits
<leader>w{H,J,K,L} will now move cursor in the corresponding direction
and exchange the currently open buffer with the one in the target
window.

Currently do not know of a way to see if there is actually a valid
window in that direction, so though this seems to behave well in the
intended scenario, doing this in any other situation is essentially
undefined behavior for now.

TODO: see if there is a way to query the geometry of splits in vim